third_party/WebKit/LayoutTests/fast/viewport/viewport-41-expected.txt

ALERT: viewport size 1000x704 scale 0.5 with limits [0.32, 5] and userScalable true
